make into one form again and handle most login in the service layer
