+++ server/src/user-org/dto/create-user-org.dto.ts
@@ -1,7 +1,7 @@
 import { IsEnum, IsNotEmpty, IsOptional } from 'class-validator';
+import { CreateOrganizationDto } from 'src/organizations/dto/create-organization.dto';
+import { CreateUserDto } from 'src/users/dto/create-user.dto';
 
-import { Organization } from '../../organizations/entities/organization.entity';
-import { User } from '../../users/entities/user.entity';
 import { ApprovalStatus, Role } from '../constants';
 
 export class CreateUserOrganizationDto {
@@ -10,12 +10,12 @@ export class CreateUserOrganizationDto {
   approvalStatus: ApprovalStatus;
 
   @IsNotEmpty({ message: 'org_id is required' })
-  organization: Organization;
+  organization: CreateOrganizationDto;
 
-  @IsNotEmpty({ message: 'role is required' })
+  @IsOptional()
   @IsEnum(Role)
   role: Role;
 
   @IsNotEmpty({ message: 'user_id is required' })
-  user: User;
+  user: CreateUserDto;
 }
