+++ server/src/user-org/user-org.module.ts
@@ -1,13 +1,17 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
+import { Organization } from 'src/organizations/entities/organization.entity';
+import { OrganizationsService } from 'src/organizations/organizations.service';
+import { User } from 'src/users/entities/user.entity';
+import { UsersService } from 'src/users/users.service';
 import { UserOrganization } from './entities/user-org.entity';
 import { UserOrganizationsController } from './user-org.controller';
 import { UserOrganizationsService } from './user-org.service';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([UserOrganization])],
+  imports: [TypeOrmModule.forFeature([UserOrganization, User, Organization])],
   controllers: [UserOrganizationsController],
-  providers: [UserOrganizationsService],
+  providers: [UserOrganizationsService, UsersService, OrganizationsService],
   exports: [UserOrganizationsService],
 })
 export class UserOrganizationsModule {}
